+++ app.go
@@ -575,7 +575,7 @@ func (a *App) ResetTraffic(id string) error {
 	return nil
 }
 
-var appVersion = "1.1.1"
+var appVersion = "1.1.2"
 
 func (a *App) GetAppInfo() AppInfoDTO {
 	return AppInfoDTO{